Separated vaccinations and NPIs into their own tabs
