+++ DashboardPages/scenarioParameters.py
@@ -18,8 +18,9 @@ from ClientResources.SharedResources import maxScenarios
 from ParameterTabs.communityParams import buildCommunityTab
 from ParameterTabs.diseaseParams import buildDiseaseTab
 from ParameterTabs.dynamicParams import buildDynamicTab
+from ParameterTabs.npiParams import buildNPITab
 from ParameterTabs.templateParams import buildTemplateTab
-from ParameterTabs.vaccinationNPIParams import buildVaccinationNPITab
+from ParameterTabs.vaccinationParams import buildVaccinationTab
 
 # Logging
 scenarioLog = logging.getLogger(__name__)
@@ -209,14 +210,16 @@ in tables and graphs generated by the dashboard.
                     (
                         diseaseTab,
                         communityTab,
-                        interventionTab,
+                        vaccineTab,
+                        npiTab,
                         dynamicTab,
                         templateTab,
                     ) = st.tabs(
                         [
                             ":material/coronavirus: Pathogen",
                             ":material/groups: Community",
-                            ":material/vaccines: Vaccination and NPIs",
+                            ":material/vaccines: Vaccination",
+                            ":material/medical_mask: NPIs",
                             ":material/manage_history: Dynamic",
                             ":material/list_alt: Templates",
                         ],
@@ -224,11 +227,12 @@ in tables and graphs generated by the dashboard.
                         key=f"paramTabs{scenarioID}",
                     )
                 else:
-                    diseaseTab, communityTab, interventionTab, templateTab = st.tabs(
+                    diseaseTab, communityTab, vaccineTab, npiTab, templateTab = st.tabs(
                         [
                             ":material/coronavirus: Pathogen",
                             ":material/groups: Community",
-                            ":material/vaccines: Vaccination and NPIs",
+                            ":material/vaccines: Vaccination",
+                            ":material/medical_mask: NPIs",
                             ":material/list_alt: Templates",
                         ],
                         on_change="rerun",
@@ -240,7 +244,10 @@ in tables and graphs generated by the dashboard.
                 if communityTab.open:
                     with communityTab:
                         buildCommunityTab(scenarioID, showAdvanced)
-                if interventionTab.open:
+                if vaccineTab.open:
+                    with vaccineTab:
+                        buildVaccinationTab(scenarioID, showAdvanced)
+                if npiTab.open:
                     containersToOpen |= {
                         f"npiContainer{scenarioID}",
                         f"schoolClosureContainer{scenarioID}",
@@ -248,8 +255,8 @@ in tables and graphs generated by the dashboard.
                         f"workGroupContainer{scenarioID}",
                         f"bccContainer{scenarioID}",
                     }
-                    with interventionTab:
-                        buildVaccinationNPITab(scenarioID, showAdvanced)
+                    with npiTab:
+                        buildNPITab(scenarioID, showAdvanced)
                 if showAdvanced and dynamicTab.open:  # type: ignore
                     with dynamicTab:  # type: ignore
                         buildDynamicTab(scenarioID)
